Refactoring: mixedClusterVersion error fix + move Stage to ReplicationState
