+++ server/src/test/java/org/opensearch/recovery/RecoveriesCollectionTests.java
@@ -114,7 +114,7 @@ public class RecoveriesCollectionTests extends OpenSearchIndexLevelReplicationTe
             final long recoveryId = startRecovery(collection, shards.getPrimaryNode(), shards.addReplica());
             final long recoveryId2 = startRecovery(collection, shards.getPrimaryNode(), shards.addReplica());
             try (RecoveriesCollection.RecoveryRef recoveryRef = collection.getRecovery(recoveryId)) {
-                ShardId shardId = recoveryRef.get().shardId();
+                ShardId shardId = recoveryRef.get().indexShard().shardId();
                 assertTrue("failed to cancel recoveries", collection.cancelRecoveriesForShard(shardId, "test"));
                 assertThat("all recoveries should be cancelled", collection.size(), equalTo(0));
             } finally {
